uszipcode
will error if you use name and country. 

also, only works for the united states with zip code, because it retrieves state name. you should add in the instructions that, if you use zip code, it works for the united states only. if you use name and country, you can do global, except the location might be the wrong location with the same name. 

In the code, you should figure out a way to tell whether the "location" input is city name and country or zip code. maybe simply check for a comma.